fix fixtures
