+++ tests/Fixture/EventsFixture.php
@@ -20,6 +20,7 @@ class EventsFixture extends TestFixture
         'id' => ['type' => 'integer'],
         'title' => ['type' => 'string', 'null' => true],
         'description' => 'text',
+        'published' => ['type' => 'string', 'length' => 1, 'default' => 'N'],
         '_constraints' => ['primary' => ['type' => 'primary', 'columns' => ['id']]]
     ];
     // @codingStandardsIgnoreEnd
@@ -33,16 +34,24 @@ class EventsFixture extends TestFixture
         [
             'id' => 1,
             'title' => 'Lorem ipsum dolor sit amet',
+            'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat.',
+            'published' => 'Y',
+        ],
+        [
+            'id' => 2,
+            'title' => 'Second event',
+            'description' => 'Second event description.',
+            'published' => 'Y'
+        ],
+        [
+            'id' => 3,
+            'title' => 'Lorem ipsum dolor sit amet',
             'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. 
 Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget \'sollicitudin\' venenatis cum 
 nullam, vivamus ut a sed, mollitia lectus. 
 Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus 
 duis vestibulum nunc mattis convallis.',
-        ],
-        [
-            'id' => 2,
-            'title' => 'Lorem ipsum dolor sit amet',
-            'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat.',
+            'published' => 'Y'
         ],
     ];
 }
