+++ tests/comparisons/Seeds/pgsql/testWithData.php
@@ -2,9 +2,9 @@
 use Migrations\AbstractSeed;
 
 /**
- * Users seed.
+ * Events seed.
  */
-class UsersSeed extends AbstractSeed
+class EventsSeed extends AbstractSeed
 {
     /**
      * Run Method.
@@ -21,35 +21,29 @@ class UsersSeed extends AbstractSeed
         $data = [
             [
                 'id' => 1,
-                'username' => 'mariano',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
-                'created' => '2007-03-17 01:16:23',
-                'updated' => '2007-03-17 01:18:31',
+                'title' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat.',
+                'published' => 'Y',
             ],
             [
                 'id' => 2,
-                'username' => 'nate',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
-                'created' => '2008-03-17 01:18:23',
-                'updated' => '2008-03-17 01:20:31',
+                'title' => 'Second event',
+                'description' => 'Second event description.',
+                'published' => 'Y',
             ],
             [
                 'id' => 3,
-                'username' => 'larry',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
-                'created' => '2010-05-10 01:20:23',
-                'updated' => '2010-05-10 01:22:31',
-            ],
-            [
-                'id' => 4,
-                'username' => 'garrett',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
-                'created' => '2012-06-10 01:22:23',
-                'updated' => '2012-06-12 01:24:31',
+                'title' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. 
+Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget \'sollicitudin\' venenatis cum 
+nullam, vivamus ut a sed, mollitia lectus. 
+Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus 
+duis vestibulum nunc mattis convallis.',
+                'published' => 'Y',
             ],
         ];
 
-        $table = $this->table('users');
+        $table = $this->table('events');
         $table->insert($data)->save();
     }
 }
