+++ tests/comparisons/Seeds/testWithDataAndFields.php
@@ -2,9 +2,9 @@
 use Migrations\AbstractSeed;
 
 /**
- * Users seed.
+ * Events seed.
  */
-class UsersSeed extends AbstractSeed
+class EventsSeed extends AbstractSeed
 {
     /**
      * Run Method.
@@ -20,24 +20,24 @@ class UsersSeed extends AbstractSeed
     {
         $data = [
             [
-                'username' => 'mariano',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
+                'title' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat.',
             ],
             [
-                'username' => 'nate',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
+                'title' => 'Second event',
+                'description' => 'Second event description.',
             ],
             [
-                'username' => 'larry',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
-            ],
-            [
-                'username' => 'garrett',
-                'password' => '$2a$10$u05j8FjsvLBNdfhBhc21LOuVMpzpabVXQ9OpC2wO3pSO0q6t7HHMO',
+                'title' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. 
+Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget \'sollicitudin\' venenatis cum 
+nullam, vivamus ut a sed, mollitia lectus. 
+Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus 
+duis vestibulum nunc mattis convallis.',
             ],
         ];
 
-        $table = $this->table('users');
+        $table = $this->table('events');
         $table->insert($data)->save();
     }
 }
